feat: add web share support
More info: https://www.w3.org/TR/web-share/
To prevent abuse, the API can only be triggered in response to some user action (such as a click event). This is why it can't be a part of the exposed bridge methods. These can be called in other ways other than click events, so there is a fallback that will cover these cases